Added OCmod unistall bug fix
https://github.com/opencart/opencart/pull/9416/files

+++ upload/admin/controller/marketplace/install.php
@@ -484,10 +484,10 @@ class ControllerMarketplaceInstall extends Controller {
 					rsort($files);
 
 					foreach ($files as $file) {
-						if (is_file($file)) {
-							unlink($file);
-						} elseif (is_dir($file)) {
-							rmdir($file);
+						if (is_dir($file)) {
+							if ($this->isDirEmpty($file)) {
+								rmdir($file);
+							}
 						}
 					}
 
@@ -496,7 +496,9 @@ class ControllerMarketplaceInstall extends Controller {
 					}
 		
 					if (is_dir($source)) {
-						rmdir($source);
+						if ($this->isDirEmpty($source)) {
+							rmdir($source);
+						}
 					}
 				}
